fix(multipart file uploader): using shared http client
this resolves an issue where the HTTP client was shared and the one client was making all the requests concurrently.
The built-in http client does not support this and resulted in interlaced writes, but it did sometimes succeed. The specs succeed as they never blocked the reactor and hence never actually executed in parallel.

I've also added configuration to limit the number of parallel uploads as previously it was uploading the complete file at once which could exhaust memory.